base class for nwb apis
