+++ allensdk/brain_observatory/nwb/nwb_api.py
@@ -1,0 +1,26 @@
+import pynwb
+
+
+class NwbApi:
+
+    __slots__ = ('path', '_nwbfile')
+
+    @property
+    def nwbfile(self):
+        if hasattr(self, '_nwbfile'):
+            return self._nwbfile
+
+        io = pynwb.NWBHDF5IO(self.path, 'r')
+        return io.read()
+
+    def __init__(self, path, **kwargs):
+        ''' Reads data for a single Extracellular Electrophysiology session from an NWB 2.0 file
+        '''
+
+        self.path = path
+
+    @classmethod
+    def from_nwbfile(cls, nwbfile, **kwargs):
+        obj = cls(path=None, **kwargs)
+        obj._nwbfile = nwbfile
+        return obj
